… config/config.exs
… mix.lock