[Card](https://rentgroup.atlassian.net/browse/SRV-5981)

This PR was branched off of https://github.com/rentpath/ex_json_logger/pull/24 in order build on what was contributed there.

* Switch to iolist to improvement performance
* Update to latest version of Erlang 26
* Minor performance tweaks
* Document benchee benchmark
* Add Elixir 1.15 and Erlang 26, remove Elixir 1.13 from GitHub workflow

## Benchmarks

<details><summary>Before</summary>

```
Operating System: macOS
CPU Information: Apple M1 Pro
Number of Available Cores: 10
Available memory: 32 GB
Elixir 1.15.7
Erlang 26.2

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: debug, info
Estimated total run time: 14 s

Benchmarking format/4 with input debug ...
Benchmarking format/4 with input info ...

##### With input debug #####
Name               ips        average  deviation         median         99th %
format/4      579.53 K        1.73 μs  ±1081.85%        1.38 μs        2.92 μs

##### With input info #####
Name               ips        average  deviation         median         99th %
format/4      377.40 K        2.65 μs   ±388.45%        2.33 μs        5.08 μs
```
</details>


<details><summary>After</summary>

```
Compiling 1 file (.ex)
Operating System: macOS
CPU Information: Apple M1 Pro
Number of Available Cores: 10
Available memory: 32 GB
Elixir 1.15.7
Erlang 26.2

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: debug, info
Estimated total run time: 14 s

Benchmarking format/4 with input debug ...
Benchmarking format/4 with input info ...

##### With input debug #####
Name               ips        average  deviation         median         99th %
format/4        1.55 M      643.44 ns  ±2151.32%         583 ns         750 ns

##### With input info #####
Name               ips        average  deviation         median         99th %
format/4      991.21 K        1.01 μs  ±1398.08%        0.92 μs        1.13 μs
```
</details>