@Stratus3D 2023-12-14
> Some failing unit tests

I fixed the tests locally. CI workflow is running tests on all versions of Erlang.